requirements.txt
google-cloud-language
protobuf
fastapi
uvicorn[standard]